Resolve type warning

+++ src/blas1/complex.cpp
@@ -112,7 +112,7 @@ void BLAS1_complex_std_vector(benchmark::State& state) {
 #ifdef XBENCHMARK_USE_XTENSOR
 template <typename T, typename Op>
 void BLAS1_complex_xarray(benchmark::State& state) {
-	const int vector_size = state.range(0);
+	const unsigned long vector_size = state.range(0);
 	T a = static_cast<T>(2.0) ;
 	xt::xarray<T> vec1 = xt::xarray<T>::from_shape({vector_size});
 	xt::xarray<T> vec2 = xt::xarray<T>::from_shape({vector_size});
@@ -138,7 +138,7 @@ void BLAS1_complex_xarray(benchmark::State& state) {
 #ifdef XBENCHMARK_USE_XTENSOR
 template <typename T, typename Op>
 void BLAS1_complex_xtensor(benchmark::State& state) {
-	const int vector_size = state.range(0);
+	const unsigned long vector_size = state.range(0);
 	T a = static_cast<T>(2.0) ;    
 	xt::xtensor<T, 1> vec1   = xt::xtensor<T,1>::from_shape({vector_size});
 	xt::xtensor<T, 1> vec2   = xt::xtensor<T,1>::from_shape({vector_size});
@@ -163,7 +163,7 @@ void BLAS1_complex_xtensor(benchmark::State& state) {
 #ifdef XBENCHMARK_USE_XTENSOR
 template <typename T, typename Op>
 void BLAS1_complex_xtensor_eval(benchmark::State& state) {
-	const int vector_size = state.range(0);
+	const unsigned long vector_size = state.range(0);
 	T a = static_cast<T>(2.0) ;    
 	xt::xtensor<T, 1> vec1   = xt::xtensor<T,1>::from_shape({vector_size});
 	xt::xtensor<T, 1> vec2   = xt::xtensor<T,1>::from_shape({vector_size});
@@ -188,7 +188,7 @@ void BLAS1_complex_xtensor_eval(benchmark::State& state) {
 #ifdef XBENCHMARK_USE_XTENSOR
 template <typename T, typename Op>
 void BLAS1_complex_xtensor_auto_eval(benchmark::State& state) {
-	const int vector_size = state.range(0);
+	const unsigned long vector_size = state.range(0);
 	T a = static_cast<T>(2.0) ;
 	xt::xtensor<T, 1> vec1   = xt::xtensor<T,1>::from_shape({vector_size});
 	xt::xtensor<T, 1> vec2   = xt::xtensor<T,1>::from_shape({vector_size});
@@ -215,7 +215,7 @@ void BLAS1_complex_xtensor_auto_eval(benchmark::State& state) {
 #ifdef XBENCHMARK_USE_XTENSOR
 template <typename T, typename Op>
 void BLAS1_complex_xtensor_only_auto_eval(benchmark::State& state) {
-	const int vector_size = state.range(0);
+	const unsigned long vector_size = state.range(0);
 	T a = static_cast<T>(2.0) ;
 	xt::xtensor<T, 1> vec1   = xt::xtensor<T,1>::from_shape({vector_size});
 	xt::xtensor<T, 1> vec2   = xt::xtensor<T,1>::from_shape({vector_size});
