+++ src/view/view_all.cpp
@@ -123,7 +123,7 @@ void VIEW_all_aligned_masked(benchmark::State& state) {
 #ifdef XBENCHMARK_USE_XTENSOR
 template <typename T>
 void VIEW_all_xarray(benchmark::State& state) {
-	const int vector_size = state.range(0);
+	const unsigned long vector_size = state.range(0);
 	xt::xarray<T> vec1 = xt::xarray<T>::from_shape({vector_size});
 	xt::xarray<T> vec2 = xt::xarray<T>::from_shape({vector_size});
 	xt::xarray<T> result = xt::xarray<T>::from_shape({vector_size});
@@ -145,7 +145,7 @@ void VIEW_all_xarray(benchmark::State& state) {
 #ifdef XBENCHMARK_USE_XTENSOR
 template <typename T>
 void VIEW_all_xtensor(benchmark::State& state) {
-	const int vector_size = state.range(0);
+	const unsigned long vector_size = state.range(0);
 	xt::xtensor<T, 1> vec1 = xt::xtensor<T, 1>::from_shape({vector_size});
 	xt::xtensor<T, 1> vec2 = xt::xtensor<T, 1>::from_shape({vector_size});
 	xt::xtensor<T, 1> result = xt::xtensor<T, 1>::from_shape({vector_size});
@@ -168,7 +168,7 @@ void VIEW_all_xtensor(benchmark::State& state) {
 #ifdef XBENCHMARK_USE_XTENSOR
 template <typename T>
 void VIEW_all_xtensor_strided(benchmark::State& state) {
-	const int vector_size = state.range(0);
+	const unsigned long vector_size = state.range(0);
 	xt::xtensor<T, 1> vec1 = xt::xtensor<T, 1>::from_shape({vector_size});
 	xt::xtensor<T, 1> vec2 = xt::xtensor<T, 1>::from_shape({vector_size});
 	xt::xtensor<T, 1> result = xt::xtensor<T, 1>::from_shape({vector_size});
@@ -189,7 +189,7 @@ void VIEW_all_xtensor_strided(benchmark::State& state) {
 #ifdef XBENCHMARK_USE_XTENSOR
 template <typename T>
 void VIEW_all_xtensor_strided_range(benchmark::State& state) {
-	const int vector_size = state.range(0);
+	const unsigned long vector_size = state.range(0);
 	xt::xtensor<T, 1> vec1 = xt::xtensor<T, 1>::from_shape({vector_size});
 	xt::xtensor<T, 1> vec2 = xt::xtensor<T, 1>::from_shape({vector_size});
 	xt::xtensor<T, 1> result = xt::xtensor<T, 1>::from_shape({vector_size});
@@ -215,7 +215,7 @@ void VIEW_all_xtensor_strided_range(benchmark::State& state) {
 // Observation : very slow !  We should try with raw pointers to compare potential performances. 
 template <typename T>
 void VIEW_all_xtensor_masked(benchmark::State& state) {
-	const int vector_size = state.range(0);
+	const unsigned long vector_size = state.range(0);
 	xt::xtensor<T, 1> vec1 = xt::xtensor<T, 1>::from_shape({vector_size});
 	xt::xtensor<T, 1> vec2 = xt::xtensor<T, 1>::from_shape({vector_size});
 	xt::xtensor<T, 1> result = xt::xtensor<T, 1>::from_shape({vector_size});
@@ -242,7 +242,7 @@ void VIEW_all_xtensor_masked(benchmark::State& state) {
 // Observation : very slow !  We should try with raw pointers to compare potential performances. 
 template <typename T>
 void VIEW_all_xtensor_masked_2(benchmark::State& state) {
-	const int vector_size = state.range(0);
+	const unsigned long vector_size = state.range(0);
 	xt::xtensor<T, 1> vec1 = xt::xtensor<T, 1>::from_shape({vector_size});
 	xt::xtensor<T, 1> vec2 = xt::xtensor<T, 1>::from_shape({vector_size});
 	xt::xtensor<T, 1> result = xt::xtensor<T, 1>::from_shape({vector_size});
@@ -266,7 +266,7 @@ void VIEW_all_xtensor_masked_2(benchmark::State& state) {
 // Observation : very slow !  We should try with raw pointers to compare potential performances. 
 template <typename T>
 void VIEW_all_xtensor_raw_masked(benchmark::State& state) {
-	const int vector_size = state.range(0);
+	const unsigned long vector_size = state.range(0);
 	xt::xtensor<T, 1> vec1 = xt::xtensor<T, 1>::from_shape({vector_size});
 	xt::xtensor<T, 1> vec2 = xt::xtensor<T, 1>::from_shape({vector_size});
 	xt::xtensor<T, 1> result = xt::xtensor<T, 1>::from_shape({vector_size});
